Support for VGM loop-offset
Vgm files can contain a loop-offset, so that when looping the song the
once-off intro can be skipped.

This patch marks the index in our simplified format where that loop offset
occurs, along with the end-index of the tune. These are then used to loop
when playing it back rather than going on to play garbage memory.